Session优化
