+++ Masuit.Tools.Core/Win32/Windows.cs
@@ -1,9 +1,7 @@
-﻿using Masuit.Tools.Hardware;
-using System;
+﻿using System;
 using System.Diagnostics;
 using System.Management;
 using System.Net;
-using System.Runtime.InteropServices;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading;
@@ -16,54 +14,6 @@ namespace Masuit.Tools.Win32
     /// </summary>
     public static class Windows
     {
-        /// <summary>
-        /// 跨平台调用C++的方法
-        /// </summary>
-        /// <param name="hwProc">程序句柄</param>
-        /// <returns></returns>
-        [DllImport("psapi.dll")]
-        static extern int EmptyWorkingSet(IntPtr hwProc);
-
-        /// <summary>
-        /// 清理系统内存，返回优化内存后的内存占用率
-        /// </summary>
-        /// <returns>优化内存后的内存占用率</returns>
-        public static double ClearMemory()
-        {
-            ClearMemorySilent();
-            Thread.Sleep(1000);
-            return SystemInfo.GetRamInfo().MemoryUsage;
-        }
-
-        /// <summary>
-        /// 静默清理系统内存
-        /// </summary>
-        public static void ClearMemorySilent()
-        {
-            GC.Collect();
-            GC.WaitForPendingFinalizers();
-            foreach (Process p in Process.GetProcesses())
-            {
-                using (p)
-                {
-                    if ((p.ProcessName.Equals("System")) && (p.ProcessName.Equals("Idle")))
-                    {
-                        //两个系统的关键进程，不整理
-                        continue;
-                    }
-
-                    try
-                    {
-                        EmptyWorkingSet(p.Handle);
-                    }
-                    catch
-                    {
-                        // ignored
-                    }
-                }
-            }
-        }
-
         /// <summary>  
         /// 获取当前使用的IP  
         /// </summary>
